requirements.txt
Flask-Cors
Flask
mysql-connector-python